Consolidating scheduler into a single module
